// src/app/models/IBriefProfile.ts
export interface IBriefProfile {
    usi: number;
    uniqueId: string;
    personTypeId: number;
    firstName: string;
    middleName?: any;
    lastSurname: string;
    name: string;
    email: string;
    role: string;
    languageCode?: any;
    imageUrl: string;
    feedbackExternalUrl: string;
}